split load shedders, add cli, request type
