+++ Cargo.toml
@@ -16,6 +16,8 @@ axum = "0.7"
 backoff = { version = "0.4", features = ["tokio"] }
 base64 = "0.22"
 bytes = "1.6"
+clap = { version = "4.5", features = ["derive", "string", "env"] }
+clap_derive = "4.5"
 chacha20poly1305 = "0.10"
 cloudflare = { git = "https://github.com/cloudflare/cloudflare-rs.git", rev = "f14720e42184ee176a97676e85ef2d2d85bc3aae", default-features = false, features = [
     "rustls-tls",
@@ -33,6 +35,7 @@ hickory-resolver = { version = "0.24", features = [
 http = "1.1"
 http-body = "1.0"
 http-body-util = "0.1"
+humantime = "2.1"
 hyper = "1.4"
 hyper-util = { version = "0.1", features = ["full"] }
 instant-acme = { version = "0.7.1", default-features = false, features = [
